untangle and remove custom oidc scenario
